+++ src/elias_fano_list.rs
@@ -1,14 +1,11 @@
 //! Compressed integer list with prefix-summed Elias-Fano encoding.
 #![cfg(target_pointer_width = "64")]
 
-pub mod iter;
-
 use std::io::{Read, Write};
 
 use anyhow::{anyhow, Result};
 use num_traits::ToPrimitive;
 
-use crate::elias_fano_list::iter::Iter;
 use crate::{EliasFano, EliasFanoBuilder, IntGetter, Searial};
 
 /// Compressed integer list with prefix-summed Elias-Fano encoding.
@@ -178,6 +175,39 @@ impl Searial for EliasFanoList {
 
     fn size_in_bytes(&self) -> usize {
         self.ef.size_in_bytes()
+    }
+}
+
+/// Iterator for enumerating integers, created by [`EliasFanoList::iter()`].
+pub struct Iter<'a> {
+    efl: &'a EliasFanoList,
+    pos: usize,
+}
+
+impl<'a> Iter<'a> {
+    /// Creates a new iterator.
+    pub const fn new(efl: &'a EliasFanoList) -> Self {
+        Self { efl, pos: 0 }
+    }
+}
+
+impl<'a> Iterator for Iter<'a> {
+    type Item = usize;
+
+    #[inline(always)]
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.pos < self.efl.len() {
+            let x = self.efl.get_int(self.pos).unwrap();
+            self.pos += 1;
+            Some(x)
+        } else {
+            None
+        }
+    }
+
+    #[inline(always)]
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        (self.efl.len(), Some(self.efl.len()))
     }
 }
